+++ pos_multi_session/__manifest__.py
@@ -7,9 +7,9 @@
     "name": """Sync POS orders across multiple sessions""",
     "summary": """Use multiple POS for handling orders""",
     "category": "Point Of Sale",
-    "live_test_url": 'http://apps.it-projects.info/shop/product/pos-multi-session?version=11.0',
+    # "live_test_url": 'http://apps.it-projects.info/shop/product/pos-multi-session?version=11.0',
     "images": ["images/pos-multi-session.png"],
-    "version": "11.0.4.2.5",
+    "version": "11.0.4.2.7",
     "application": False,
 
     "author": "IT-Projects LLC, Ivan Yelizariev",