+++ pos_multi_session_sync/__manifest__.py
@@ -5,9 +5,9 @@
     "name": """Sync Server for POS orders""",
     "summary": """Sync orders via main or separate (e.g. local) server""",
     "category": "Point of Sale",
-    "live_test_url": 'http://apps.it-projects.info/shop/product/pos-multi-session?version=11.0',
+    # "live_test_url": 'http://apps.it-projects.info/shop/product/pos-multi-session?version=11.0',
     "images": [],
-    "version": "11.0.1.0.3",
+    "version": "11.0.1.0.4",
     "application": False,
 
     "author": "IT-Projects LLC, Kolushov Alexandr",