+++ pos_orders_history/__manifest__.py
@@ -7,9 +7,9 @@
     "name": "POS Orders History",
     "summary": """See all paid orders from special menu in POS""",
     "category": "Point of Sale",
-    "live_test_url": "http://apps.it-projects.info/shop/product/pos-orders-history?version=11.0",
+    # "live_test_url": "http://apps.it-projects.info/shop/product/pos-orders-history?version=11.0",
     "images": ['images/pos_orders_history_main.png'],
-    "version": "11.0.1.2.0",
+    "version": "11.0.1.2.1",
     "application": False,
 
     "author": "IT-Projects LLC, Dinar Gabbasov",