+++ pos_pin/static/src/js/pos.js
@@ -1,3 +1,6 @@
+/*  Copyright 2016 Krotov Stanislav <https://github.com/ufaks>
+    Copyright 2018-2019 Kolushov Alexandr <https://it-projects.info/team/KolushovAlexandr>
+    License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl.html).*/
 odoo.define('pos_pin.pos', function (require) {
     "use strict";
 
@@ -8,6 +11,7 @@ odoo.define('pos_pin.pos', function (require) {
 
     gui.Gui.include({
         sudo_custom: function(options) {
+            options = options || {};
             var user = options.user || this.pos.get_cashier();
         
             if ($.inArray(options.special_group, user.groups_id) >= 0) {
@@ -22,7 +26,7 @@ odoo.define('pos_pin.pos', function (require) {
         select_user_custom: function(options){
             options = options || {};
             var self = this;
-            var def  = new $.Deferred();
+            var def = new $.Deferred();
 
             var list = [];
             for (var i = 0; i < this.pos.users.length; i++) {
@@ -38,21 +42,28 @@ odoo.define('pos_pin.pos', function (require) {
             this.show_popup('selection',{
                 'title': options.title || _t('Select User'),
                 'list': list,
-                'confirm': function(user){ def.resolve(user); },
-                'cancel':  function(){ def.reject(); }
-            });
-
-            return def.then(function(user){
-                if (options.security && user !== options.current_user && user.pos_security_pin) {
-                    return self.ask_password(user.pos_security_pin, options.arguments).then(function(){
-                        return self.set_and_render_cashier(user);
-                    });
-                } else {
-                    return self.set_and_render_cashier(user);
+                'confirm': function(cashier){
+                    def.resolve(cashier);
+                },
+                'cancel':  function(){
+                    def.reject();
                 }
             });
+
+            return def.then(function(cashier){
+                if (options.security && cashier !== options.current_user && cashier.pos_security_pin) {
+                    return self.ask_password(cashier.pos_security_pin, options.arguments);
+                }
+                return cashier;
+            }).done(function(res){
+                return self.check_then_set_and_render_cashier(options, res);
+            });
         },
-        set_and_render_cashier: function(user){
+
+        check_then_set_and_render_cashier: function(options, user) {
+            if (options.do_not_change_cashier){
+                return user;
+            }
             if (this.pos.get_cashier().id !== user.id) {
                 this.pos.set_cashier(user);
                 this.pos.chrome.widget.username.renderElement();
@@ -76,7 +87,9 @@ odoo.define('pos_pin.pos', function (require) {
                     }
                 },
                 cancel: function() {
-                    cancel_function.call(self);
+                    if (cancel_function) {
+                        cancel_function.call(self);
+                    }
                     lock.reject();
                 },
             });