+++ pos_cashier_select/doc/changelog.rst
@@ -1,5 +1,10 @@
+`1.0.1`
+-------
+
+**Fix:** Incompatibility with pos_product_category_discount module
+
 `1.0.0`
 -------
 
-- Init version
+**Init version**