+++ pos_multi_session/doc/changelog.rst
@@ -1,3 +1,14 @@
+`4.2.7`
+-------
+
+- **Fix:** The Multi-sessions menu is not available for POS Managers
+- **Fix:** Error on using different companies within the same multi-session
+
+`4.2.6`
+-------
+
+- **Fix:** Unsynced order in some cases
+
 `4.2.5`
 -------